requirements.txt
# Web Framework
fastapi==0.104.1
uvicorn==0.23.2
pydantic==2.4.2
python-dotenv==1.0.0
python-multipart==0.0.6

# RAGFlow e OCR
ragflow==1.0.3
deepdocs==0.9.2
pillow==10.0.1
pytesseract==0.3.10

# Embeddings e LLM
openai==1.3.0
tiktoken==0.5.1

# FAISS
faiss-cpu==1.7.4  # Use faiss-gpu para ambientes com GPU

# Processamento de dados
numpy==1.24.4
pandas==2.1.1
scikit-learn==1.3.2
scipy==1.11.3

# Visualização
matplotlib==3.8.1
plotly==5.18.0

# Utilitários
httpx==0.25.1
tenacity==8.2.3
cachetools==5.3.1
loguru==0.7.2
ujson==5.8.0
python-jose[cryptography]==3.3.0  # Para JWT se implementar autenticação

# Desenvolvimento e testes
pytest==7.4.3
pytest-asyncio==0.21.1
httpx==0.25.1
black==23.10.1
isort==5.12.0
mypy==1.6.1